refactor: move debug command
